Move getInboxUnseenCount to internal services/inboxBadge module
Moves getInboxUnseenCount from the exported extension entrypoint to
an internal module (services/inboxBadge.ts) so it is not part of the
public API surface exposed via coreExtension.exports.
